Split the E2E tests into smoke/cassandra/es
