+++ CONTRIBUTING.adoc
@@ -180,3 +180,36 @@ OLM Integration:
 
 Total Score: 4/18 points
 ----
+
+==== E2E tests
+
+The whole set of end-to-end tests can be executed via:
+
+[source,bash]
+----
+$ make e2e-tests
+----
+
+The end-to-end tests are split into tags and can be executed in separate groups, such as:
+
+[source,bash]
+----
+$ make e2e-tests-smoke
+----
+
+Other targets include `e2e-tests-cassandra` and `e2e-tests-elasticsearch`. Refer to the `Makefile` for an up-to-date list of targets.
+
+If you face issues like the one below, make sure you don't have any Jaeger instances (`kubectl get jaegers`) running nor Ingresses (`kubectl get ingresses`):
+
+[source]
+----
+--- FAIL: TestSmoke (316.59s)
+    --- FAIL: TestSmoke/smoke (316.55s)
+        --- FAIL: TestSmoke/smoke/daemonset (115.54s)
+...
+...
+            daemonset.go:30: timed out waiting for the condition
+...
+...
+----
+
