+++ test/e2e/suite_elasticsearch_test.go
@@ -1,0 +1,29 @@
+// +build elasticsearch
+
+package e2e
+
+import (
+	"testing"
+
+	framework "github.com/operator-framework/operator-sdk/pkg/test"
+	"github.com/stretchr/testify/assert"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
+	"github.com/jaegertracing/jaeger-operator/pkg/apis"
+	"github.com/jaegertracing/jaeger-operator/pkg/apis/jaegertracing/v1"
+)
+
+func TestElasticsearch(t *testing.T) {
+	assert.NoError(t, framework.AddToFrameworkScheme(apis.AddToScheme, &v1.JaegerList{
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "Jaeger",
+			APIVersion: "jaegertracing.io/v1",
+		},
+	}))
+
+	t.Run("elasticsearch", func(t *testing.T) {
+		t.Run("spark-dependencies-es", SparkDependenciesElasticsearch)
+		t.Run("simple-prod", SimpleProd)
+		// t.Run("es-index-cleaner", EsIndexCleaner)
+	})
+}
